refactor: Separate FolderMenu component
